refactor: extract profile layout mutations hook from MainLayout
Made-with: Cursor
